+++ CHANGELOG.md
@@ -2,5 +2,7 @@
 
 **Experimental. Do not use in production.**
 
+* Add `retire-users` command to `tutor local` to run the retirement pipeline
+* Add retirement-job service
 * Add basic testing via tox
 * Initial Git import